core: fix import template from ova
When importing a template from OVA we don't go through
ImportVmTemplateFromConfiguration and therefore
getImageToAvailableStorageDomains() returns an empty map and later we
fail with NPE.

This patch changes the fix for https://bugzilla.redhat.com/2043124 to be
specific for ImportVmTemplateFromConfiguration and so other
import-template flows would not face this issue.

Bug-Url: https://bugzilla.redhat.com/2074916
